DateFormatter class upgraded to DateTimeUtils, will all required construction, formating and calculation methods

+++ src/main/java/pk/gov/pbs/utils/SystemUtils.java
@@ -14,7 +14,7 @@ import java.io.UnsupportedEncodingException;
 
 public class SystemUtils {
     public static long getUnixTs(){
-        return (long)(System.currentTimeMillis() / 1000);
+        return System.currentTimeMillis() / 1000;
     }
 
     public static boolean isNetworkAvailable(Context context) {
